backend/pyproject.toml
[project]
name = "photosite-backend"
version = "0.1.0"
description = "the manifest generating backend for my personal photography website"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "fsspec>=2025.9.0",
    "pillow>=11.3.0",
    "pyexiftool>=0.5.6",
    "s3fs>=0.4.2",
    "typer>=0.19.1",
    "ujson>=5.11.0",
]

[project.scripts]
main = "photosite_backend.main:app"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.4.0",
    "pyrefly>=0.31.1",
    "pytest>=8.4.1",
    "pytest-pretty>=1.3.0",
    "ruff>=0.12.10",
    "uv>=0.8.13",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--color=yes"
]

[tool.ruff.lint]
extend-select = [
    # no prints in code
    "T201",
    # no fstring in logger
    "G004"
]
